Valid Blocks Count
You are given an integer array A of length N. where each element is between 2 and 10
(inclusive). You have to count how many consecutive blocks of identical elements exist
such that the length of the block is equal to the value of the element. Your task is to
find and return an integer value representing the count of such blocks.

Input Specification:
input1 : An integer value N representing the size of the array.
input2 : An integer array A.

Output Specification:

Return an integer representing the count of such blocks that match the above
condition.

Example 1:

input1 : 11
input2 : (2,3,3,3,2,2,6,4,4,4,4)