Create knight problem
In an Arrayland. N knights line up in a row A, each with a unique power level. A knight ii considered special if the number of stronger knights to their left is strictly greafer than the number of stronger knights to their right. Your task is to find and return an integer value representing the count of special knights in the lineup. Note: A stronger knight is a knight with higher power level. Input Specification: input1 : An integer value N, representing the number of knights. input2 : An integer array A, representing the power level of the knights. utput Specification: Return an integer value representing the number of special knights in the lineup.