Create Double array
input1 : {1,2,1} input2 : 3 In a bustling tech company, a data scientist is tasked with preparing a dataset for analysis. She needs to create a "Doubled Dataset" to conduct art various experiments. This dataset should consist of the original list of user data followed by an exact duplicate to ensure that her models have enough data for testing. You are given the dataset in the form on and array A, your task is to return an array which consists of the original array followed by the same array. hib2 Output : {1,2,1,1,2,1} parth Input Specification: input1 : An integer array A representing the dataset. input2 : An integer value representing the length of the array. Explanation: Here, the given array is {1,2,1}. When the same elements are added to the original array the array becomes {1,2,1,1,2,1} therefore, {1,2,1,1,2,1} is returned as the output. Example 2: input1 : {1,3,2,1} input2 : 4 Output : {1,3,2,1,1,3,2,1} Output Specification: Return an array which consist of the original array followed by the same array. il.cor Explanation: Here, the given array is {1,3,2,1}. When the same elements are added to the original array the array becomes {1,3,2,1,1,3,2,1} therefore, {1,3,2,1,1,3,2,1} is returned as the output.