Create sum of digits
In a small village, the townsfolk discovered a number N written on an ancient scroll. They believe that to unlock its secrets, they must reduce the ar number N to a single digit by summing its digits repeatedly. Your task is to help the townsfolk to unlock the scroll by returning an integer value representing the reduced single digit of the number N. Input Specification: input1 : An integer value representing the number N written on the ancient scroll 62 Example 1: input1 : 38 urthib2" part Output Specification: Return an integer value representing the reduced single digit of the number N. Output : 2 Explanation: The number on the ancient scroll is 38, to reduce this number we will add the digits 3 + 8 = 11. Since the number is not a single digit number, the digits will be added again. Now the number becomes 1 + 1 = 2, which is a single digit number. Therefore, 2 is returned as the output.