Create Knowledge Enhancement
In a magical library, each bookshelf is represented by a two-dimensional array A, where each row of the 2D array A[i] represents the series value of a book A row is considered magical if the sum of the odd values of the series of a book is even. aYour task is to find and return an integer value representing the number of magical rows. Input Specification: input1 : An integer value representing the number of rows in the 2D array. input2 : An integer value representing the number of columns in the 2D array. input3 : A 2D integer array where each row represents a series of books. Output Specification: Return an integer value representing the number of magical rows. Output : 2 vittay Explanation: Here, the given 2D array is ({1, 2, 3), (4, 5, 6), {7, 8, 9}], 8 - In the first row (1, 2, 3), the odd numbers are {1, 3} and their sum is - In the second row {4, 5, 6}, the odd numbers are {5} and as there is element so the sum is 5 which is odd. - In the third row (7, 8, 9}, the odd numbers are (7, 9} and their sum even. Therefore, there are only 2 magical rows so, 2 is returned as the outp Example 2: input1 :3 input2 : 2 input3 : ({2, 4], {0,0}, {11, 11]} Example 1: input1: 3 input2 : 3 input3 : ((1, 2, 3), (4, 5, 6), {7, 8, 9]]