House Visit
You are given an integer array A of size N. which represents houses. From i1 house, you can go directly to (i+A[i))th house, if it is present. Your task is to find and return an integer value representing the total number of houses you can visit while traveling from the first house to the last. Example 1: input1 :4 input2 : (2,1,3,1) Output : 2 Note: · Assume 1 based indexing. . We alays have to count the fust house. Input Specification: input1 : An integer value N, representing the size of the array input2 : An integer array A of positive integers, representing the houses. Output Specification: Return an integer value representing the total number of houses you can visit while traveling from the first house to the last. xample 1: input1: 4 input2 : (2,1,3,1) 4019 Explanation: Here the given array is (2,1,3,1). The houses can be visited in the following manner: . The first house that is visited will be at index 1 with value 2. . The next house to be visited from the 15 house will be at index 3 (1+2) with value of 3. · The next house to be visited from the 310 house will be at index 6 (3+3) which is not present. Since no more houses are present, a total of 2 houses can be visited. Therefore, 2 is a oftretumed as the output. Example 2: input1: 5 input2 : (1,3,5,2,8} utput : 2 Output : 3